requirements.txt
pandas
numpy
faiss-cpu
sentence-transformers
tqdm
scikit-learn
streamlit
rapidfuzz
